preprocess/data_process_0/requirements.txt
langchain-openai>=0.1.0
langgraph>=0.1.0
pandas>=2.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0